Use previous pthread stack address range to avoid looping through /proc/self/maps.
Android frequently loops throgh /proc/self/maps to determine the stack
size of current pthreads process'. However, when a process has many
mappings this task can become tedious and a lot of CPU time can be
is wasted in finding stack size multiple times.

To avoid this we can utilise the fact that the stack address range
is not cleared after being calculated, and avoid searching for it
again when the current PID has already done so. In case the stack
address range has been modified, we invalidate the "cache" and allow
getstack to search /proc/self/maps again.

Suggested-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>